fix error in async_sglang_server when finding workers
