+++ verl/workers/rollout/sglang_rollout/async_sglang_server.py
@@ -47,15 +47,17 @@ class AsyncSglangServer(AsyncServerBase):
             actor for actor in all_actors if actor.get("name", None).startswith(self.wg_prefix + "WorkerDict_")
         ]
 
+        gpu_per_node = len(set([actor["name"].split(":")[1] for actor in matched_actors]))
+
         for matched_actor in matched_actors:
             fields = matched_actor["name"].split(":")
             assert len(fields) == 2, f"invalid actor name: {matched_actor['name']}"
             pg_index, local_rank = int(fields[0].split("_")[-1]), int(fields[1])
 
-            if (self._dp_size * pg_index + local_rank) // self._tp_size == self._dp_rank:
+            if (gpu_per_node * pg_index + local_rank) // self._tp_size == self._dp_rank:
                 worker = ray.get_actor(**matched_actor)
                 self.workers.append(worker)
-                if (self._dp_size * pg_index + local_rank) / self._tp_size == self._dp_rank:
+                if (gpu_per_node * pg_index + local_rank) / self._tp_size == self._dp_rank:
                     self.master_worker = worker
 
     async def chat_completion(self, raw_request: Request):
